backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
lightgbm==4.5.0
shap==0.46.0
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.5.0
faker==28.0.0
python-multipart==0.0.9
python-dotenv==1.0.0
joblib==1.4.0
psutil==5.9.8
google-genai==1.14.0
